+++ apps/cowswap-frontend/src/modules/swap/containers/BottomBanners/index.tsx
@@ -4,6 +4,7 @@ import HAND_SVG from '@cowprotocol/assets/cow-swap/hand.svg'
 import { BannerOrientation, InlineBanner, StatusColorVariant } from '@cowprotocol/ui'
 import { useWalletInfo } from '@cowprotocol/wallet'
 
+import { Trans } from '@lingui/react/macro'
 import { Link } from 'react-router'
 import styled from 'styled-components/macro'
 import { WIDGET_MAX_WIDTH } from 'theme'
@@ -39,7 +40,9 @@ export function BottomBanners(): ReactNode {
           backDropBlur
           margin="10px auto auto"
         >
-          Funds stuck? <Link to={accountProxyUrl}>Recover your funds</Link>
+          <Trans>
+            Funds stuck? <Link to={accountProxyUrl}>Recover your funds</Link>
+          </Trans>
         </InlineBanner>
       )}
     </Wrapper>