+++ apps/cowswap-frontend/src/modules/swap/containers/SwapWidget/index.tsx
@@ -4,6 +4,8 @@ import { useFeatureFlags } from '@cowprotocol/common-hooks'
 import { isInjectedWidget, isSellOrder } from '@cowprotocol/common-utils'
 import { useIsEagerConnectInProgress, useIsSmartContractWallet, useWalletInfo } from '@cowprotocol/wallet'
 
+import { t } from '@lingui/core/macro'
+
 import { Field } from 'legacy/state/types'
 import { useHooksEnabledManager } from 'legacy/state/user/hooks'
 
@@ -130,18 +132,19 @@ export function SwapWidget({ topContent, bottomContent }: SwapWidgetProps): Reac
     fiatAmount: outputCurrencyFiatAmount,
     receiveAmountInfo: isSellTrade ? receiveAmountInfo : null,
   }
+
   const inputCurrencyPreviewInfo = {
     amount: inputCurrencyInfo.amount,
     fiatAmount: inputCurrencyInfo.fiatAmount,
     balance: inputCurrencyInfo.balance,
-    label: isSellTrade ? 'Sell amount' : 'Expected sell amount',
+    label: isSellTrade ? t`Sell amount` : t`Expected sell amount`,
   }
 
   const outputCurrencyPreviewInfo = {
     amount: outputCurrencyInfo.amount,
     fiatAmount: outputCurrencyInfo.fiatAmount,
     balance: outputCurrencyInfo.balance,
-    label: isSellTrade ? 'Receive (before fees)' : 'Buy exactly',
+    label: isSellTrade ? t`Receive (before fees)` : t`Buy exactly`,
   }
 
   const rateInfoParams = useRateInfoParams(inputCurrencyInfo.amount, outputCurrencyInfo.amount)
@@ -190,7 +193,7 @@ export function SwapWidget({ topContent, bottomContent }: SwapWidgetProps): Reac
           <>
             {bottomContent}
             <SwapRateDetails rateInfoParams={rateInfoParams} deadline={deadlineState[0]} />
-            {isPrimaryValidationPassed && <TradeApproveWithAffectedOrderList />}
+            {isPrimaryValidationPassed && isPartialApproveEnabled && <TradeApproveWithAffectedOrderList />}
             <Warnings buyingFiatAmount={buyingFiatAmount} />
             {tradeWarnings}
             <TradeButtons
@@ -215,6 +218,7 @@ export function SwapWidget({ topContent, bottomContent }: SwapWidgetProps): Reac
         toBeImported,
         intermediateBuyToken,
         isPrimaryValidationPassed,
+        isPartialApproveEnabled,
       ],
     ),
   }