+++ apps/cowswap-frontend/src/pages/Hooks/index.tsx
@@ -1,5 +1,7 @@
 import { PAGE_TITLES } from '@cowprotocol/common-const'
 
+import { useLingui } from '@lingui/react/macro'
+
 import { PageTitle } from 'modules/application/containers/PageTitle'
 import { HooksStoreWidget } from 'modules/hooksStore'
 import { SwapUpdaters } from 'modules/swap'
@@ -7,9 +9,11 @@ import { SwapUpdaters } from 'modules/swap'
 // TODO: Add proper return type annotation
 // eslint-disable-next-line @typescript-eslint/explicit-function-return-type
 export function HooksPage() {
+  const { i18n } = useLingui()
+
   return (
     <>
-      <PageTitle title={PAGE_TITLES.HOOKS} />
+      <PageTitle title={i18n._(PAGE_TITLES.HOOKS)} />
       <SwapUpdaters />
       <HooksStoreWidget />
     </>