+++ libs/hook-dapp-lib/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## [1.13.0](https://github.com/cowprotocol/cowswap/compare/hook-dapp-lib-v1.12.0...hook-dapp-lib-v1.13.0) (2025-11-05)
+
+
+### Features
+
+* **hook-dapps:** register aave flash-loans hooks ([#6470](https://github.com/cowprotocol/cowswap/issues/6470)) ([a033304](https://github.com/cowprotocol/cowswap/commit/a033304a8a6376ccb28319e998c13a7916eebcd5))
+
+
+### Bug Fixes
+
+* update permit hook description ([#6412](https://github.com/cowprotocol/cowswap/issues/6412)) ([0342200](https://github.com/cowprotocol/cowswap/commit/03422007648aede61f75ac8b66b450e2a01bbe8a))
+
 ## [1.12.0](https://github.com/cowprotocol/cowswap/compare/hook-dapp-lib-v1.11.0...hook-dapp-lib-v1.12.0) (2025-10-02)